requirements.txt
aws-cdk-lib==2.37.1
constructs>=10.0.0,<11.0.0
PyYAML
requests